Add publisher module
This implements the publisher integration (https://github.com/honeycombio/agentless-integrations-for-aws) used to asynchronously forward events from Lambda to Honeycomb. We use it internally, but it's a bit of tech debt for us because we use the CFN template to deploy it while we use terraform everywhere else, so this would be nice to have.

First time writing one of these modules for publishing. I copied this to our internal modules directory and ran tf plan - it did everything I was expecting, but curious if I missed anything.